# Add udivrem3by2 and use it in Knuth division