… div_reciprocalTable.go