Includes #39.

The `udivrem3by2()` is very useful to implement `udivremBy2` procedure similar to `udivremBy1` one (https://github.com/holiman/uint256/blob/master/uint256.go#L395). This is very helpful for the performance of division by 128-bit numbers (worst case for Knuth division as the size difference between numerator and divisor is the biggest). However, this is not done in this PR.

Having `udivrem3by2()` implemented, it can be also used in Knuth directly. The Knuth algorithm is effectively doing the same, to this is mostly code cleanup and reuse. This has been done in intx: https://github.com/chfast/intx/pull/134.

Here, in current form this actually brings performance regressions: most importantly for cases `/mod129` and `/mod256`.

```
name                     old time/op  new time/op   delta
Div/small/uint256-6      11.4ns ± 0%   11.7ns ± 0%   +2.63%  (p=0.000 n=20+19)
Div/mod64/uint256-6      62.6ns ± 0%   53.1ns ± 0%  -15.18%  (p=0.000 n=16+20)
Div/mod128/uint256-6     97.2ns ± 0%  109.5ns ± 0%  +12.61%  (p=0.000 n=17+20)
Div/mod192/uint256-6     87.6ns ± 0%   95.7ns ± 0%   +9.28%  (p=0.000 n=13+20)
Div/mod256/uint256-6     73.5ns ± 0%   80.5ns ± 0%   +9.56%  (p=0.000 n=20+19)
Mod/small/uint256-6      13.7ns ± 0%   14.0ns ± 0%   +2.19%  (p=0.000 n=19+20)
Mod/mod64/uint256-6      65.1ns ± 0%   55.3ns ± 0%  -15.05%  (p=0.000 n=20+20)
Mod/mod128/uint256-6      100ns ± 0%    112ns ± 0%  +12.00%  (p=0.000 n=20+17)
Mod/mod192/uint256-6     90.0ns ± 0%   97.7ns ± 0%   +8.58%  (p=0.000 n=18+19)
Mod/mod256/uint256-6     76.5ns ± 0%   83.1ns ± 0%   +8.71%  (p=0.000 n=20+19)
AddMod/small/uint256-6   18.4ns ± 0%   18.6ns ± 0%   +1.09%  (p=0.000 n=20+17)
AddMod/mod64/uint256-6   73.7ns ± 0%   63.6ns ± 0%  -13.67%  (p=0.000 n=19+18)
AddMod/mod128/uint256-6   116ns ± 0%    129ns ± 0%  +11.21%  (p=0.000 n=20+19)
AddMod/mod192/uint256-6   107ns ± 0%    116ns ± 0%   +8.41%  (p=0.000 n=20+20)
AddMod/mod256/uint256-6  95.0ns ± 0%  102.0ns ± 0%   +7.34%  (p=0.000 n=18+19)
MulMod/small/uint256-6   37.0ns ± 0%   36.7ns ± 0%   -0.81%  (p=0.000 n=20+18)
MulMod/mod64/uint256-6    106ns ± 0%     96ns ± 0%   -9.06%  (p=0.000 n=20+14)
MulMod/mod128/uint256-6   186ns ± 0%    213ns ± 0%  +14.52%  (p=0.000 n=20+19)
MulMod/mod192/uint256-6   183ns ± 0%    205ns ± 0%  +12.02%  (p=0.000 n=20+17)
MulMod/mod256/uint256-6   178ns ± 0%    196ns ± 0%  +10.11%  (p=0.000 n=15+19)
_SDiv/large/uint256-6     101ns ± 0%    104ns ± 0%   +2.97%  (p=0.000 n=16+19)
```